+++ FE/src/components/common/QuestionSelectionBox/AnswerSelectionModal/AnswerSelectionModal.tsx
@@ -2,10 +2,10 @@ import Modal from '@foundation/Modal';
 import { css } from '@emotion/react';
 import { theme } from '@styles/theme';
 import Typography from '@foundation/Typography/Typography';
-import useQuestionAnswerQuery from '@hooks/queries/useQuestionAnswerQuery';
+import useQuestionAnswerQuery from '@/hooks/apis/queries/useQuestionAnswerQuery';
 import AnswerScript from './AnswerScript';
 import AnswerForm from './AnswerForm';
-import useAnswerDefaultMutation from '@hooks/mutations/useAnswerDefaultMutation';
+import useAnswerDefaultMutation from '@/hooks/apis/mutations/useAnswerDefaultMutation';
 import { Question } from '@/types/question';
 
 type AnswerSelectionModalProps = {
