+++ FE/src/components/common/QuestionSelectionBox/QuestionTabPanelItem.tsx
@@ -7,7 +7,7 @@ import { css } from '@emotion/react';
 import { useState } from 'react';
 import { useRecoilValue } from 'recoil';
 import QuestionAccordion from './QuestionAccordion';
-import useQuestionCategoryQuery from '@hooks/queries/useQuestionCategoryQuery';
+import useQuestionCategoryQuery from '@/hooks/apis/queries/useQuestionCategoryQuery';
 import Toggle from '@foundation/Toggle/Toggle';
 import QuestionAddForm from './QuestionAddForm';
