+++ FE/src/components/interviewVideoPage/Modal/VideoShareModal.tsx
@@ -3,7 +3,7 @@ import Typography from '@foundation/Typography/Typography';
 import { css } from '@emotion/react';
 import ShareRangeSetting from '@components/interviewVideoPage/Modal/ShareRangeSetting';
 import VideoShareModalFooter from '@components/interviewVideoPage/Modal/VideoShareModalFooter';
-import useToggleVideoPublicMutation from '@hooks/mutations/video/useToggleVideoPublicMutation';
+import useToggleVideoPublicMutation from '@/hooks/apis/mutations/useToggleVideoPublicMutation';
 import { truncateText } from '@/utils/textUtils';
 
 type VideoShareModalProps = {
