+++ FE/src/hooks/apis/mutations/useDeleteVideoMutation.ts
